Sound and Difficulty
Difficulty has been updated for levels 5-8. The levels have been altered to require 2-3 transformations for early levels and 3-4 for later ones. Certain levels may need to be rearranged.

UI sound assets were changed to have decrease volume. The sound sources for the AI voice in all levels was lowered to .25.